Change the int32 mirror lag field to int64 (#11)
Change the int32 mirror lag field to int64

+++ kafkarestv3/model_mirror_lag.go
@@ -11,5 +11,5 @@ package kafkarestv3
 // MirrorLag struct for MirrorLag
 type MirrorLag struct {
 	Partition int32 `json:"partition"`
-	Lag int32 `json:"lag"`
+	Lag int64 `json:"lag"`
 }
